+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+### Upcoming
+- Remove `add(subviews: [VFLView], names: [String])`
+- Add `updateConstraints()`
+- Add ` appendConstraints([NSLayoutConstraint])`
+- Simplify examples
+
 ### 1.0.0
 - First draft
 - Add `add(subview: name:)` variants to subviews